requirements.txt
# LS API 자동매매 시스템 패키지 목록

# 이베스트 API 및 Windows COM
pywin32>=306

# 데이터베이스
mysql-connector-python>=8.0.0
PyMySQL>=1.0.0

# 환경설정
python-dotenv>=1.0.0

# 데이터 처리
numpy>=1.24.0
pandas>=2.0.0

# 텔레그램 봇
python-telegram-bot>=20.0

# 날짜/시간 처리
python-dateutil>=2.8.0

# 웹 요청
requests>=2.28.0

# 로깅
PyYAML>=6.0

# 테스트
pytest>=7.0.0
pytest-cov>=4.0.0

# 기타 유틸리티
urllib3>=1.26.0

# 금융 데이터 (옵션)
# pykrx>=1.3.0
# FinanceDataReader>=0.9.0